made changes to repository

+++ app/src/main/java/no/uio/ifi/in2000/team19/prosjekt/model/DTO/locationForecast/Geometry.kt
@@ -1,6 +1,6 @@
 package no.uio.ifi.in2000.team19.prosjekt.model.DTO.locationForecast
 
 data class Geometry(
-    val coordinates: List<Int>,
+    val coordinates: List<Double>,
     val type: String
 )
